session state y_colum structure updated
st.session_state.y_columns stores all the graph data needed. Name, absolute data, relative data, if the graph is visible and the graph color. Previously it was a list of list, so you needed to use y_columns[i][0] to access the name or y_columns[i][3] to access the relative data. Because this isn't readable, you can now access it with y_columns[i].name or y_columns[i].relativeData